requirements.txt
gradio
aiogram
python-dotenv
groq
loguru